+++ src/GLib/GLib.stub.php
@@ -247,6 +247,13 @@ final class GKeyFile
     public function get_boolean(string $group_name, string $key): bool {}
 
     /**
+     * Returns the values associated with $key under $group_name as booleans.
+     *
+     * @return list<bool>
+     */
+    public function get_boolean_list(string $group_name, string $key): array {}
+
+    /**
      * Retrieves a comment above $key from $group_name. If $key is `null` then $comment will be
      * read from above $group_name. If both $key and $group_name are `null`, then $comment will be
      * read from above the first group in the file.
@@ -260,6 +267,21 @@ final class GKeyFile
     public function get_double(string $group_name, string $key): float {}
 
     /**
+     * Returns the values associated with $key under $group_name as doubles.
+     *
+     * @return list<float>
+     */
+    public function get_double_list(string $group_name, string $key): array {}
+
+    /**
+     * Returns all groups in the key file loaded with $key_file. The array of returned groups will
+     * be `null`-terminated, so $length may optionally be `null`.
+     *
+     * @return list<string>
+     */
+    public function get_groups(): array {}
+
+    /**
      * Returns the value associated with $key under $group_name as a signed 64-bit integer. This is
      * similar to g_key_file_get_integer() but can return 64-bit results without truncation.
      */
@@ -267,6 +289,22 @@ final class GKeyFile
 
     /** Returns the value associated with $key under $group_name as an integer. */
     public function get_integer(string $group_name, string $key): int {}
+
+    /**
+     * Returns the values associated with $key under $group_name as integers.
+     *
+     * @return list<int>
+     */
+    public function get_integer_list(string $group_name, string $key): array {}
+
+    /**
+     * Returns all keys for the group name $group_name. The array of returned keys will be
+     * `null`-terminated, so $length may optionally be `null`. In the event that the $group_name
+     * cannot be found, `null` is returned and $error is set to %G_KEY_FILE_ERROR_GROUP_NOT_FOUND.
+     *
+     * @return list<string>
+     */
+    public function get_keys(string $group_name): array {}
 
     /**
      * Returns the actual locale which the result of g_key_file_get_locale_string() or
@@ -280,6 +318,14 @@ final class GKeyFile
      */
     public function get_locale_string(string $group_name, string $key, ?string $locale): string {}
 
+    /**
+     * Returns the values associated with $key under $group_name translated in the given $locale if
+     * available. If $locale is `null` then the current locale is assumed.
+     *
+     * @return list<string>
+     */
+    public function get_locale_string_list(string $group_name, string $key, ?string $locale): array {}
+
     /** Returns the name of the start group of the file. */
     public function get_start_group(): ?string {}
 
@@ -288,6 +334,13 @@ final class GKeyFile
      * g_key_file_get_value(), this function handles escape sequences like \s.
      */
     public function get_string(string $group_name, string $key): string {}
+
+    /**
+     * Returns the values associated with $key under $group_name.
+     *
+     * @return list<string>
+     */
+    public function get_string_list(string $group_name, string $key): array {}
 
     /**
      * Returns the value associated with $key under $group_name as an unsigned 64-bit integer. This
@@ -407,6 +460,9 @@ final class GKeyFile
 
     /** Associates a new value with $key under $group_name. */
     public function set_value(string $group_name, string $key, string $value): void {}
+
+    /** This function outputs $key_file as a string. */
+    public function to_data(): string {}
 
     public static function error_quark(): int {}
 }